requirements.txt
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
numpy
scipy
scikit-learn
matplotlib
pandas
seaborn
joblib
pickle = 3.0
pprint
re
os
tqdm
word_tokenize
stopwords
WordNetLemmatizer
RegexpTokenizer
shuffle
LabelEncoder
train_test_split
tensorflow-cpu
Adam
relu
Tokenizer
load_model
pad_sequences
Input
Dropout
Embedding
BatchNormalization
LSTM
Dense
GRU
Bidirectional
CORS
json